requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
openai
passlib
python-jose
bcrypt
pydantic
email-validator
python-multipart
loguru